feat(*): fix Drawer layout
