+++ packages/plasma-b2c/src/components/Drawer/Drawer.component-test.tsx
@@ -43,7 +43,7 @@ describe('plasma-b2c: Drawer', () => {
         const [isOpen, setIsOpen] = React.useState(false);
         const {
             width = '50vw',
-            height = '100vh',
+            height,
             placement = 'right',
             closePlacement = 'right',
             hasClose = true,
