+++ packages/plasma-new-hope/src/components/Popup/ui/index.ts
@@ -1,0 +1,2 @@
+export { Draggable } from './Draggable/Draggable';
+export { Resizable } from './Resizable/Resizable';
