+++ packages/plasma-web/src/components/Drawer/Drawer.component-test.tsx
@@ -53,7 +53,7 @@ describe('plasma-b2c: Drawer', () => {
         const [isOpen, setIsOpen] = React.useState(false);
         const {
             width = '50vw',
-            height = '100vh',
+            height,
             placement = 'right',
             closePlacement = 'right',
             hasClose = true,
