+++ packages/sdds-cs/src/components/Drawer/Drawer.update-test.component-test.tsx
@@ -54,7 +54,7 @@ describe('sdds-cs: Drawer', () => {
         const [isOpen, setIsOpen] = React.useState(false);
         const {
             width = '50vw',
-            height = '100vh',
+            height,
             placement = 'right',
             closePlacement = 'right',
             hasClose = true,
